requirements.txt
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
ciscoisesdk==2.4.0
colorama==0.4.6
coverage==7.12.0
cryptography==46.0.3
dict2xml==1.7.7
dill==0.4.0
distlib==0.4.0
f5-icontrol-rest==1.3.13
fastjsonschema==2.21.2
filelock==3.20.0
genie==25.10
genie.libs.clean==25.10
genie.libs.conf==25.10
genie.libs.filetransferutils==25.10
genie.libs.health==25.10
genie.libs.ops==25.10
genie.libs.parser==25.10
genie.libs.sdk==25.10
grpcio==1.76.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
invoke==2.2.1
Jinja2==3.1.6
jsonpickle==4.1.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ncclient==0.7.0
netaddr==1.3.0
netmiko==4.6.0
nodeenv==1.9.1
nornir==3.5.0
nornir-netmiko==1.0.1
nornir-utils==0.2.0
ntc_templates==8.1.0
packaging==25.0
paramiko==4.0.0
platformdirs==4.5.0
pluggy==1.6.0
ply==3.11
pre_commit==4.5.0
prettytable==3.17.0
protobuf==6.33.1
pyasn1==0.6.0
pyasynchat==1.0.4
pyasyncore==1.0.4
pycparser==2.23
pyftpdlib==2.1.0
Pygments==2.19.2
PyNaCl==1.6.1
pyserial==3.5
pysmi==1.6.2
pysnmp==6.1.4
pysnmpcrypto==0.0.4
pytest==9.0.1
pytest-cov==4.1.0
pyvmomi==9.0.0.0
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rest.connector==25.10
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scp==0.15.0
setuptools==80.9.0
textfsm==2.1.0
tftpy==0.8.0
tqdm==4.67.1
typing_extensions==4.15.0
unicon==25.10
unicon.plugins==25.10
urllib3==2.5.0
virtualenv==20.35.4
wcwidth==0.2.14
wheel==0.45.1
xmltodict==0.12.0
yang.connector==25.10
